fix: code after review
