+++ src/components/Main/index.jsx
@@ -1,25 +1,10 @@
-import React, { useRef } from 'react';
-import Promo from './Promo';
-import AboutProject from './AboutProject';
-import Tech from './Techs';
-import AboutMe from './AboutMe';
-import Portfolio from './Portfolio';
+import React from 'react';
 import './Main.css';
 
-const Main = () => {
-  const ref = useRef();
-
-  const scroll = () => window.scrollTo({ top: ref.current.offsetTop, behavior: 'smooth' });
-
-  return (
+const Main = ( props ) => (
     <main className="main">
-      <Promo scroll={scroll}/>
-      <AboutProject ref={ref}/>
-      <Tech />
-      <AboutMe />
-      <Portfolio />
+      { props.children }
     </main>
-  );
-};
+);
 
 export default Main;
