+++ src/components/MovieCardList/MoviesCardList.css
@@ -4,7 +4,11 @@
   flex-direction: column;
   align-items: center;
   width: 100%;
-  padding: 84px 0 0;
+  padding: 70px 0 0;
+}
+
+.movies-list__saved {
+  padding-bottom: 120px;
 }
 
 .movies-list:before {
@@ -28,7 +32,7 @@
   border: 1px solid #E8E8E8;
   border-radius: 3px;
   background-color: inherit;
-  margin: 40px auto 80px;
+  margin: 50px auto 80px;
   transition: opacity .5s;
   font-size: 12px;
   line-height: 15px;
@@ -43,19 +47,27 @@
     width: calc(100% - 60px);
   }
 
+  .movies-list__saved {
+    padding-bottom: 120px;
+  }
+
   .movies-list__container {
     width: calc(100% - 60px);
     grid-template-columns: 1fr 1fr;
   }
 
   .movies-list__yet {
-    margin-top: 30px;
+    margin-top: 35px;
   }
 }
 
 @media screen and (max-width: 470px) {
   .movies-list {
-    padding-top: 64px;
+    padding-top: 55px;
+  }
+
+  .movies-list__saved {
+    padding-bottom: 140px;
   }
 
   .movies-list:before {
@@ -69,6 +81,6 @@
 
   .movies-list__yet {
     width: 240px;
-    margin: 50px 0 80px;
+    margin: 65px 0 80px;
   }
 }
