+++ src/components/Navigation/Navigation.css
@@ -36,6 +36,8 @@
   display: none;
   background-image: url("../../images/burger.svg");
   background-size: cover;
+  border: none;
+  background-color: inherit;
   width: 44px;
   height: 44px;
   margin: 15px 30px;
@@ -48,6 +50,8 @@
   right: 22px;
   background-image: url("../../images/close.svg");
   background-size: cover;
+  border: none;
+  background-color: inherit;
   width: 32px;
   height: 32px;
 }
