+++ src/components/Promo/Promo.css
@@ -52,7 +52,7 @@
   height: 304px;
 }
 
-@media screen and (max-width: 768px) {
+@media screen and (max-width: 800px) {
   .promo {
     display: flex;
     flex-direction: column;
