+++ src/components/Techs/Tech.css
@@ -52,15 +52,17 @@
   border: none;
   border-radius: 10px;
   background-color: #E8E8E8;
-  flex-grow: 1;
   margin: 0 5px;
   max-width: 90px;
   text-align: center;
+  flex-grow: 1;
+  flex-shrink: 1;
+  flex-basis: 90px;
 }
 
-@media screen and (max-width: 768px) {
+@media screen and (max-width: 800px) {
   .tech {
-    padding: 87px 50px;
+    padding: 90px 50px;
   }
 
   .tech__container {
@@ -75,10 +77,25 @@
   .tech__technology {
     font-size: 12px;
     line-height: 14px;
+    max-width: 90px;
   }
 
   .tech__subtitle {
     padding: 0 0 22px;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  .tech__technologies {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr 1fr;
+    gap: 10px;
+    max-width: 400px;
+    margin: 0 auto;
+  }
+
+  .tech__technology {
+    margin: 0;
   }
 }
