+++ src/components/Techs/index.jsx
